% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFSMOutput.R
\name{PlotFSMOutput}
\alias{PlotFSMOutput}
\title{Plot FSM Ouput}
\usage{
PlotFSMOutput(output, machine.index = 1, suppress.lower = 0)
}
\arguments{
\item{output}{Output (list) of a previous call to RunFSMSyntax.}

\item{machine.index}{Machine that should be plotted}

\item{suppress.lower}{Supress path in transition frequency plot with relative frequency below that threshold}
}
\description{
Under Development
}
