% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareLogData.R
\name{PrepareLogData}
\alias{PrepareLogData}
\title{Prepare Log Data form Different Fromats for Use with LogFSM}
\usage{
PrepareLogData(zipfilename, outfilename, workingdir, verbose = F,
  elements = "", datafiletype = "dataflatv01a", flags = "", ...)
}
\arguments{
\item{zipfilename}{Name of the file containing the raw log data that should be used with LogFSM. 'PrepareLogData' will read from this file.}

\item{outfilename}{Name of the file in which 'PrepareLogData' will write the prepared log data.}

\item{workingdir}{Working directory (optional).}

\item{verbose}{Requesting more detailed output from LogFSM.}

\item{elements}{A string variable refering to the element names (i.e., items, units or tasks), that should be extracted. Multiple elemtns can be combinded using ';'.}

\item{datafiletype}{Type of the data provided in the file \code{zipfilename} (either \code{dataflatv01a} or \code{piaaczip01a})}

\item{flags}{Optional flags as documented for the specific data formats. The flag "RELATIVETIME" can be used to prepare log data provided with relative timestamps. Multiple flages can be combined in the string variable using the pipe (|).}

\item{...}{(Further arguments will be passed on if necessary)}
}
\value{
The function returns TRUE, if a file was created.
}
\description{
Differtent data sources can be used with the LogFSM package. In order to harmonize the processing, data
must be converted in the internal LogFSM format (\code{logfsmjson}) using this function. The function must be
called with the filename (\code{zipfilename}) in which the data are stored and a filename (\code{outfilename})
in which the prepared data are written.
}
\details{
Data formats that require pre-processing using the function \code{PrepareLogData}
to create the \code{logfsmjson}(i.e. data that can be used with \code{RunFSMSyntax})
\itemize{
\item \code{dataflatv01a}: Flat table with log data
\item \code{piaaczip01a}: Text file with log data from the Programme for the International Assessment of Adult Competencies (PIAAC), exported from PIAAC LogDataAnalyzer (R1)
\item \code{nepszip01a}: ZIP archive with data in generic log data format used in the National Educational Panel Study (NEPS, TBT-studies 2009-2020)
}

If data are provided as  \code{dataflatv01a}, the following columns are expected:
\itemize{
\item \code{PersonIdentifier}: Identifier for the test-taker / person / student
\item \code{EventName}: Name of the log event
\item \code{Element}: Item, unit, task (part of the assessemnt) to which the event belongs
\item \code{TimeStamp} or \code{RelativeTime}: Timestamp (or relative time, if the flag \code{RELATIVETIME} is used)
}

If data are provided as  \code{nepszip01a}, the following columns are expected:
\itemize{
\item \code{PersonIdentifier}:  Identifier for the test-taker / person / student (Provide a value as argument 'ColumnNamePersonIdentifier' to override this default.)
\item \code{EventName}: Name of the log event (Provide a value as argument 'ColumnNameEventName' to override this default.)
\item \code{Element}: Item, unit, task (part of the assessemnt) to which the event belongs (Provide a value as argument 'ColumnNameElement' to override this default.)
\item \code{TimeStamp} or \code{RelativeTime}: Timestamp (Absolut time stamps are expected. For relative times use the flag \code{RELATIVETIME}. Provide a value as argument 'ColumnNameTimeStamp' to override the default column name.)
}

Additional formats (under development)
\itemize{
\item \code{pisabqzip01a} / \code{pisabqzip01b} / \code{pisabqzip01c} / \code{pisacazip01a}
}
}
