% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntheticLogData.R
\name{SyntheticLogData}
\alias{SyntheticLogData}
\title{Synthetic Data for the Ddocumentation of the LogFSM Syntax}
\usage{
SyntheticLogData(example = "ex01a")
}
\arguments{
\item{example}{Name of the example (e.g. ```ex01a````)}
}
\description{
This function provides synthetic example data for to illustrate the useage of the LogFSM syntax.
}
\details{
Example \code{ex01a} and \code{ex01b}
\itemize{
\item Simple item with 3 pages (text1, text2, question)
}

Example \code{ex02a}
\itemize{
\item Simple log data to illustrate guards.
}
}
