% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteLogData.R
\name{WriteLogData}
\alias{WriteLogData}
\title{Write Log Data}
\usage{
WriteLogData(file, data)
}
\arguments{
\item{file}{Filename}

\item{data}{Data frame with log data (\emph{Flat and sparse log data table})}
}
\description{
Writes log data from data frame \code{data} to JSON format (\code{jsonlite}), that can be used to
run a finite-state machine with the function \code{RunFSMSyntax.}
}
\details{
The following columns names are taken as keywords:
\itemize{
\item \code{PersonIdentifier} (required)
\item \code{TimeStamp} (required)
\item \code{EventName} (required)
\item \code{TimeDifferencePrevious} (optional)
}
}
